Remove ContextualConnectionPool
